+++ Source/UnrealFlecs/Components/FlecsModuleComponent.h
@@ -3,7 +3,9 @@
 #pragma once
 
 #include "CoreMinimal.h"
+
 #include "SolidMacros/Macros.h"
+
 #include "FlecsModuleComponent.generated.h"
 
 USTRUCT(BlueprintType)
